+++ CHANGELOG.md
@@ -10,6 +10,9 @@ changes.
 
 ## [0.10.0] - UNRELEASED
 
+- **LOG BREAKING** internally, the Rollback event has been changed to carry the
+  chain state which changes a bit the format of some logs.
+
 - **BREAKING** Changed `hydra-node` and `hydra-tui` arguments:
   + Instead of `--network-id` flag they now use `--testnet-magic` and we support
     also `--mainnet` flag.